Added the Unity project
Open the PhysicsHandandSoftBodySphere unity scene to see the demo I
showed Dr. Tucker.  If you run this scene, there will be a softbody
sphere, and some buttons across the top of the screen.  The buttons
instantiate more objects.  If you have a Leap Connected, the hand should
automatically appear and force data from collisions will be printed to
the console.